Add iD editor
iD sources live in vendor/assets/iD, and are managed with
vendorer, using the "rewrite" functionality to replace relative
paths in the CSS with <%= asset_path(…) %> directives.

To avoid needing to namespace all of its CSS classes, iD
is embedded via an iframe.

To setup, edit application.yml to include an `id_key` key.
(It can use the same consumer key as P2 if you like.)